% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{package}
\name{app-package}
\alias{app}
\alias{app-package}
\title{shiny app for interactive learning}
\description{
PKG_DESC.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/D-Se/app}
  \item Report bugs at \url{https://github.com/D-Se/app/issues}
}

}
\author{
\strong{Maintainer}: Donald Seinen \email{2869830202@qq.com}

}
\keyword{internal}
