% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe.R
\name{safe}
\alias{safe}
\title{Safely evaluate arbitrary R expressions}
\usage{
safe(expr)
}
\arguments{
\item{expr}{An arbitrary R expression}
}
\value{
The result of evaluating the object: for an expression vector this
 is the result of evaluating the last element.
}
\description{
Safely evaluate arbitrary R expressions
}
